requirement.txt
uvicorn==0.17.6
pymongo==4.1.1
bcrypt==3.2.2
python-multipart==0.0.5
mongoengine==0.24.1
fastapi==0.87.0
fastapi-login==1.8.2
pydantic==1.9.1
PyJWT==2.4.0
loguru==0.6.0
bcrypt==3.2.2
pytz